+++ packages/aspida/samples/v1.1.ts
@@ -1,6 +1,6 @@
 import { mockMethods } from 'aspida-mock'
 
-export interface Methods {
+export type Methods = {
   get: {
     // test
     query?: ({ aa: number /*