+++ packages/aspida/samples/v1.1/2/_hogeId@HogeId/entries.json.ts
@@ -1,4 +1,4 @@
-export interface Methods {
+export type Methods = {
   get: {
     resBody: {
       id: number