+++ packages/openapi2aspida/samples/openapi/api/v1/mixes/_id.ts
@@ -1,5 +1,5 @@
 /* eslint-disable */
-export interface Methods {
+export type Methods = {
   put: {
     reqBody?: {
       name: string