+++ packages/openapi2aspida/samples/strapi/users/index.ts
@@ -1,7 +1,7 @@
 /* eslint-disable */
 import * as Types from '../@types'
 
-export interface Methods {
+export type Methods = {
   get: {
     query?: {
       _limit?: number